+++ test/core-download/outputComplete.test.js
@@ -1,6 +1,8 @@
 const cd0 = require('../../lib/core-download')
+const cou = require('../../lib/core-output')
+const m = require('../mocklib')
 
-test('download/outputComplete', () => {
+test('download/outputComplete/1', () => {
   const fileWorkingObject = {
     fileNum: 1,
     fileSize: 1000,
@@ -9,18 +11,16 @@ test('download/outputComplete', () => {
     remoteFile: './test.txt'
   }
 
-  const logValues = [
-    ['1 downloaded ./test.txt ./test.txt 1000']
+  const expectedInfo = [
+    '1 downloaded ./test.txt ./test.txt 1000'
   ]
 
-  const logOutput = []
-  global.console.info = jest.fn().mockImplementation((s) => { logOutput.push([s]) })
+  const i = new m.MockOutput()
+  cou.info = i.fn()
 
   cd0.outputComplete(fileWorkingObject)
     .then(() => {
-      expect(console.info).toBeCalled()
-      logValues.forEach((x) => expect(logOutput).toContainEqual(x))
-      logOutput.forEach((x) => expect(logValues).toContainEqual(x))
+      expect(i.verify(expectedInfo)).resolves.toBe(true)
     })
 })
 
@@ -33,10 +33,10 @@ test('download/outputComplete/quiet', () => {
     remoteFile: './test.txt'
   }
 
-  global.console.info = jest.fn()
+  cou.info = jest.fn()
 
   cd0.outputComplete(fileWorkingObject)
     .then(() => {
-      expect(console.info).not.toBeCalled()
+      expect(cou.info).not.toBeCalled()
     })
 })