+++ test/core-options-verify/verifyLocalPath.test.js
@@ -1,9 +1,12 @@
 const util = require('util')
 const cov = require('../../lib/core-options-verify')
+const co0 = require('../../lib/core-output')
+const m = require('../mocklib')
 
 const DEFAULT_LOCALPATH = '.'
+const expectedWarn = [util.format('Warning: localPath is undefined, defaulting to current directory. [%s]', DEFAULT_LOCALPATH)]
 
-test('verifyLocalPath', () => {
+test('verifyLocalPath/text', () => {
   const workingObject = {
     userOption: { localPath: '/home/testuser/files/' },
     validatedOption: {}
@@ -23,9 +26,8 @@ test.each([null, ''])('verifyLocalPath/blank', (value) => {
     validatedOption: {}
   }
 
-  const warnOutput = []
-  global.console.warn = jest.fn().mockImplementation((s) => { warnOutput.push(s) })
-  const msg = util.format('Warning: localPath is undefined, defaulting to current directory. [%s]', DEFAULT_LOCALPATH)
+  const w = new m.MockOutput()
+  co0.warn = w.fn()
 
   expect(cov.verifyLocalPath(workingObject))
     .resolves
@@ -33,8 +35,7 @@ test.each([null, ''])('verifyLocalPath/blank', (value) => {
       userOption: { localPath: value },
       validatedOption: { localPath: DEFAULT_LOCALPATH }
     })
-  expect(console.warn).toBeCalled()
-  expect(warnOutput).toContain(msg)
+  expect(w.verify(expectedWarn)).resolves.toBe(true)
 })
 test('verifyLocalPath/undefined', () => {
   const workingObject = {
@@ -42,17 +43,15 @@ test('verifyLocalPath/undefined', () => {
     validatedOption: {}
   }
 
-  const warnOutput = []
-  global.console.warn = jest.fn().mockImplementation((s) => { warnOutput.push(s) })
-  const msg = util.format('Warning: localPath is undefined, defaulting to current directory. [%s]', DEFAULT_LOCALPATH)
+  const w = new m.MockOutput()
+  co0.warn = w.fn()
   expect(cov.verifyLocalPath(workingObject))
     .resolves
     .toMatchObject({
       userOption: {},
       validatedOption: { localPath: DEFAULT_LOCALPATH }
     })
-  expect(console.warn).toBeCalled()
-  expect(warnOutput).toContain(msg)
+  expect(w.verify(expectedWarn)).resolves.toBe(true)
 })
 
 test.each([null, ''])('verifyLocalPath/blank/quiet', (value) => {
@@ -61,7 +60,7 @@ test.each([null, ''])('verifyLocalPath/blank/quiet', (value) => {
     validatedOption: { quiet: true }
   }
 
-  global.console.warn = jest.fn()
+  co0.warn = jest.fn()
 
   expect(cov.verifyLocalPath(workingObject))
     .resolves
@@ -72,7 +71,7 @@ test.each([null, ''])('verifyLocalPath/blank/quiet', (value) => {
         quiet: true
       }
     })
-  expect(console.warn).not.toBeCalled()
+  expect(co0.warn).not.toBeCalled()
 })
 test('verifyLocalPath/undefined/quiet', () => {
   const workingObject = {
@@ -80,7 +79,7 @@ test('verifyLocalPath/undefined/quiet', () => {
     validatedOption: { quiet: true }
   }
 
-  global.console.warn = jest.fn()
+  co0.warn = jest.fn()
   expect(cov.verifyLocalPath(workingObject))
     .resolves
     .toMatchObject({
@@ -90,5 +89,5 @@ test('verifyLocalPath/undefined/quiet', () => {
         quiet: true
       }
     })
-  expect(console.warn).not.toBeCalled()
+  expect(co0.warn).not.toBeCalled()
 })