+++ test/core-options-verify/verifySkipIfExists.test.js
@@ -1,7 +1,11 @@
 const util = require('util')
 const cov = require('../../lib/core-options-verify')
+const co0 = require('../../lib/core-output')
+const m = require('../mocklib')
 
 const DEFAULT_SKIPIFEXISTS = false
+
+const expectedWarn = [util.format('Warning: skipIfExists is undefined, defaulting to %s.', DEFAULT_SKIPIFEXISTS)]
 
 test.each([true, 'Y', 'on', 1, 'y', 'yes'])('verifySkipIfExists/true', (value) => {
   const workingObject = {
@@ -36,8 +40,8 @@ test('verifySkipIfExists/undefined', () => {
     userOption: {},
     validatedOption: {}
   }
-  const warnOutput = []
-  global.console.warn = jest.fn().mockImplementation((s) => { warnOutput.push(s) })
+  const w = new m.MockOutput()
+  co0.warn = w.fn()
 
   expect(cov.verifySkipIfExists(workingObject))
     .resolves
@@ -45,17 +49,15 @@ test('verifySkipIfExists/undefined', () => {
       userOption: {},
       validatedOption: { skipIfExists: DEFAULT_SKIPIFEXISTS }
     })
-  const msg = util.format('Warning: skipIfExists is undefined, defaulting to %s.', DEFAULT_SKIPIFEXISTS)
-  expect(console.warn).toBeCalled()
-  expect(warnOutput).toContain(msg)
+  expect(w.verify(expectedWarn)).resolves.toBe(true)
 })
 test.each([null, ''])('verifySkipIfExists/blank', (value) => {
   const workingObject = {
     userOption: { skipIfExists: value },
     validatedOption: {}
   }
-  const warnOutput = []
-  global.console.warn = jest.fn().mockImplementation((s) => { warnOutput.push(s) })
+  const w = new m.MockOutput()
+  co0.warn = w.fn()
 
   expect(cov.verifySkipIfExists(workingObject))
     .resolves
@@ -63,12 +65,10 @@ test.each([null, ''])('verifySkipIfExists/blank', (value) => {
       userOption: { skipIfExists: value },
       validatedOption: { skipIfExists: DEFAULT_SKIPIFEXISTS }
     })
-  const msg = util.format('Warning: skipIfExists is undefined, defaulting to %s.', DEFAULT_SKIPIFEXISTS)
-  expect(console.warn).toBeCalled()
-  expect(warnOutput).toContain(msg)
+  expect(w.verify(expectedWarn)).resolves.toBe(true)
 })
 
-test.each(['ANC', '3453'])('verifySkipIfExists/not-boolean', (value) => {
+test.each(['ANC', '3453', 'xxx', 567, { x: 1 }])('verifySkipIfExists/not-boolean', (value) => {
   const workingObject = {
     userOption: { skipIfExists: value },
     validatedOption: {}
@@ -84,7 +84,7 @@ test('verifySkipIfExists/undefined/quiet', () => {
     userOption: {},
     validatedOption: { quiet: true }
   }
-  global.console.warn = jest.fn()
+  co0.warn = jest.fn()
 
   expect(cov.verifySkipIfExists(workingObject))
     .resolves
@@ -95,7 +95,7 @@ test('verifySkipIfExists/undefined/quiet', () => {
         quiet: true
       }
     })
-  expect(console.warn).not.toBeCalled()
+  expect(co0.warn).not.toBeCalled()
 })
 
 test.each([null, ''])('verifySkipIfExists/blank/quiet', (value) => {
@@ -103,7 +103,7 @@ test.each([null, ''])('verifySkipIfExists/blank/quiet', (value) => {
     userOption: { skipIfExists: value },
     validatedOption: { quiet: true }
   }
-  global.console.warn = jest.fn()
+  co0.warn = jest.fn()
 
   expect(cov.verifySkipIfExists(workingObject))
     .resolves
@@ -114,5 +114,5 @@ test.each([null, ''])('verifySkipIfExists/blank/quiet', (value) => {
         quiet: true
       }
     })
-  expect(console.warn).not.toBeCalled()
+  expect(co0.warn).not.toBeCalled()
 })