+++ test/core-options-verify/verifySkipIfOlder.test.js
@@ -1,7 +1,10 @@
 const util = require('util')
 const cov = require('../../lib/core-options-verify')
+const co0 = require('../../lib/core-output')
+const m = require('../mocklib')
 
 const DEFAULT_SKIPIFOLDER = false
+const expectedWarn = [util.format('Warning: skipIfOlder is undefined, defaulting to %s.', DEFAULT_SKIPIFOLDER)]
 
 test.each([true, 'Y', 'on', 1, 'y', 'yes'])('verifySkipIfOlder/true', (value) => {
   const workingObject = {
@@ -36,8 +39,8 @@ test('verifySkipIfOlder/undefined', () => {
     userOption: {},
     validatedOption: {}
   }
-  const warnOutput = []
-  global.console.warn = jest.fn().mockImplementation((s) => { warnOutput.push(s) })
+  const w = new m.MockOutput()
+  co0.warn = w.fn()
 
   expect(cov.verifySkipIfOlder(workingObject))
     .resolves
@@ -45,9 +48,7 @@ test('verifySkipIfOlder/undefined', () => {
       userOption: {},
       validatedOption: { skipIfOlder: DEFAULT_SKIPIFOLDER }
     })
-  const msg = util.format('Warning: skipIfOlder is undefined, defaulting to %s.', DEFAULT_SKIPIFOLDER)
-  expect(console.warn).toBeCalled()
-  expect(warnOutput).toContain(msg)
+  expect(w.verify(expectedWarn)).resolves.toBe(true)
 })
 
 test.each([null, ''])('verifySkipIfOlder/blank', (value) => {
@@ -55,8 +56,8 @@ test.each([null, ''])('verifySkipIfOlder/blank', (value) => {
     userOption: { skipIfOlder: value },
     validatedOption: {}
   }
-  const warnOutput = []
-  global.console.warn = jest.fn().mockImplementation((s) => { warnOutput.push(s) })
+  const w = new m.MockOutput()
+  co0.warn = w.fn()
 
   expect(cov.verifySkipIfOlder(workingObject))
     .resolves
@@ -64,9 +65,7 @@ test.each([null, ''])('verifySkipIfOlder/blank', (value) => {
       userOption: { skipIfOlder: value },
       validatedOption: { skipIfOlder: DEFAULT_SKIPIFOLDER }
     })
-  const msg = util.format('Warning: skipIfOlder is undefined, defaulting to %s.', DEFAULT_SKIPIFOLDER)
-  expect(console.warn).toBeCalled()
-  expect(warnOutput).toContain(msg)
+  expect(w.verify(expectedWarn)).resolves.toBe(true)
 })
 test.each(['ANC', '3453', 'xxx', 567])('verifySkipIfOlder/not-boolaen', (value) => {
   const workingObject = {
@@ -84,7 +83,7 @@ test('verifySkipIfOlder/undefined/quiet', () => {
     userOption: {},
     validatedOption: { quiet: true }
   }
-  global.console.warn = jest.fn()
+  co0.warn = jest.fn()
 
   expect(cov.verifySkipIfOlder(workingObject))
     .resolves
@@ -95,7 +94,7 @@ test('verifySkipIfOlder/undefined/quiet', () => {
         quiet: true
       }
     })
-  expect(console.warn).not.toBeCalled()
+  expect(co0.warn).not.toBeCalled()
 })
 
 test.each([null, ''])('verifySkipIfOlder/blank/quiet', (value) => {
@@ -103,7 +102,7 @@ test.each([null, ''])('verifySkipIfOlder/blank/quiet', (value) => {
     userOption: { skipIfOlder: value },
     validatedOption: { quiet: true }
   }
-  global.console.warn = jest.fn()
+  co0.warn = jest.fn()
 
   expect(cov.verifySkipIfOlder(workingObject))
     .resolves
@@ -114,5 +113,5 @@ test.each([null, ''])('verifySkipIfOlder/blank/quiet', (value) => {
         quiet: true
       }
     })
-  expect(console.warn).not.toBeCalled()
+  expect(co0.warn).not.toBeCalled()
 })